requirements.txt
google-adk
mcp
litellm
python-dotenv
fastapi
requests
streamlit
uvicorn[standard]